requirements.txt
djangorestframework==latest
Django~=5.0.6
django-filter==24.2